@@ -2,10 +2,8 @@ package com.next.reservations.api
 
 import com.next.reservations.web.mapper.ReservationMapper
 import com.next.reservations.web.request.ReservationRequest
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
+import com.next.reservations.web.response.ReservationResponse
+import org.springframework.web.bind.annotation.*
 
 
 @RestController
@@ -17,4 +15,18 @@ class ReservationController(private val mapper: ReservationMapper) {
         return mapper.createReservation(request)
     }
 
+    @GetMapping("/by-time-config/{configId}")
+    fun findAllActiveByConfigId(@PathVariable configId: Long): List<ReservationResponse> =
+            mapper.findAllActiveReservationsByConfig(configId)
+
+    @GetMapping("/by-reservation-time/{resTimeId}")
+    fun findAllActiveByReservationTimeId(@PathVariable resTimeId: Long): List<ReservationResponse> =
+            mapper.findAllActiveReservationsByReservationTime(resTimeId)
+
+    @GetMapping("/default")
+    fun findAllByDefaultConfig(): List<ReservationResponse> = mapper.findAllByDefaultConfig()
+
+    @GetMapping("/by-new-future-default/{newFutureDefault}")
+    fun findAllInvalidByNewFutureDefault(@PathVariable newFutureDefault: String): List<ReservationResponse> =
+            mapper.findAllInvalidByFutureDefaultDate(newFutureDefault)
 }
