@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize
 
 @JsonSerialize
 class ReservationTimeRequest (
-    val id: Long,
+    val id: Long?,
     val time: String,
     val reservationTimeConfigurationId: Long
 )
