@@ -1,3 +1,3 @@
 export enum ReservationStatus {
-  PENDING, ACCEPTED, REJECTED, REMOVED
+  PENDING, ACCEPTED, REJECTED, REMOVED, FINISHED
 }
