@@ -1,10 +1,13 @@
 export interface Reservation{
+  id: number;
   fullName: string;
   email: string;
   phoneNumber: string;
   comment: string;
-  numberOfPlayer: number;
+  numberOfPlayers: number;
   reservationTimeId: number;
   date: string;
   status: number;
+  previousReservationTimeDate: string;
+  currentReservationTimeDate: string;
 }
